Recipe creation
I upgraded the RecipeCreateView() which is located in view.py.

It gives logged-in user ability to upload their recipe(s) WITH MAX 9 categories.

The link itself is not yet included. The only way to access the view is by entering the following URL - "localhost:port/recipe/new"

Once form is successfully posted and validated, the user gets redirected to recipes-home (//localhost:port) with a successful message. If validation fails, the same process takes place but this time an error is displayed.