Searchbar functionality
User is able to search the recipe title in recipes-home template and users recipes in user-recipes template.

Additional change: The display of cookies has been modified because it was annoying how it manipulated the contents below, and now, once user accepts the cookies, this agreement will expire with in a day.