Fixed validation of RecipeUpdateVIew
I haven't noticed the issue when trying updating, creating and removing categories from recipe on a large scale.

The error was simple, the new data sent by the user was out of range because of odd removing script for global list element(s) such as query_ids, new_category and new_ingredients.

This issue is hopefully now resolved, but further testing would be needed to confirm it.

Some minor changes have been made, like, improved documentation of the RecipeUpdateView and improved name display of Category model in admin page.