Download recipe, included image
I converted the recipe image to base64 url so when user downloads the recipe, the html file will include the image.

Also, now the file is being saved as the name of the recipe.

I also adjusted the image position for the offline mode of the downloaded file, so it will now float on the right-hand side.