Official Icon
I designed and created a website icon. We can see the chef and a book. The chef highlights the user and the book presents recipes / MyBakes. Unfortunately, the book looks more like a body of the chef, but it doesn't matter that much. The most important thing is that it's unique.