Recipes contain new data
User is now able to set their recipe category/type -> a soup, bake, desert or other.

Website visitors may inspect new details which are found in recipe-detail template. History -> it tracks the changes that have been made to the recipe and also includes the origin date e.g. date first publishment of the recipe on BakeToom + creation date.

I also rename author field from Recipe class to publisher because author of the recipe may differ, if so, they will have to be included in the LICENSE of the recipe.