Functionality is added to majority of the search filters
I've included certification icon, which will be displayed in recipe detail container once recipe is certified.

There are only 2 filters, which miss functionality -> general date and publish date from search by section.

User is now able to choose their recipe nationality to international as default (which sends an empty string for the back-end), and I made so the recipes with international nationality will display the message "international/międzynarodowy" for the recipe detail container instead of the flag.