Fixes and hosting media files
Media files are now hosted via Amazon S3 (Bucket service).

I need to create a new save method for image upload. For now, images are not re-scalable.

I also fixed an issue in recipe_detail template for history data in the detailed container. The issue: Displayed history data used to mislead the visitor. Once the user updates their recipe, it will display a section which says that recipe was RECENTLY edited and uploaded in RECENT UPLOAD DATE. The visitor couldn't know when exactly was the recipe edited so I fixed this issue. However, when upload date is updated and then the recipe is posted, it will display the previous message that is was RECENTLY updated and uploaded on 'X' date, which means, edited at the same date as posted.