Upgraded from bootstrap4 to bootstrap5 and more
I decided to upgrade project to match the never evolving technology. I was able to do it as project's scale is small, therefore, if we decide to scale, I'd rather have at my disposal available tools of the future instead of complaining later down the line.

However, this decision costs a lot of nerves. Formats where broken due to integration, hence the design change in some areas of the website. For example, deletion of country picker library. I had to make my own. But I decided to keep the country icons from world-flags-sprite library.

My initial goal was to include credits page. However, due to my questionable decision, I had to stop developing the page. Therefore, I will try to work on it now as there is not much to include.

I'm also still deciding on whether or not I should remove the scroll-to-top button. So... yeay ;)